various refactorings and fixes
