+++ lib/execution/_all.ts
@@ -1,0 +1,2 @@
+export {LocalExecutionEnvironment} from './base-execution-env.js';
+export {DotnetExecutionEnvironment} from './dotnet-execution-env.js';
