+++ lib/execution/index.ts
@@ -1,0 +1,7 @@
+import {makeKeyedTypeGetter} from '../keyed-type.js';
+
+import * as all from './_all.js';
+
+export * from './_all.js';
+
+export const getExecutionEnvironmentByKey = makeKeyedTypeGetter('executionEnvironment', all);
